@@ -26,22 +26,29 @@
     "README.md",
     "UNLICENSE"
   ],
+  "scripts": {
+    "test": "ava",
+    "coverage": "nyc npm test"
+  },
+  "ava": {
+    "files": "test/**/*.test.js"
+  },
   "dependencies": {
-    "balanced-match": "^0.3.0",
+    "balanced-match": "^0.4.1",
     "postcss": "^5.0.12"
   },
   "devDependencies": {
+    "ava": "^0.15.2",
     "babel-cli": "^6.1.18",
-    "babel-eslint": "^4.1.5",
-    "babel-istanbul": "^0.6.0",
+    "babel-eslint": "^6.1.0",
     "babel-preset-es2015": "^6.1.18",
+    "babel-preset-stage-0": "^6.5.0",
+    "chai": "^3.5.0",
     "coveralls": "^2.11.2",
-    "eslint": "^1.9.0",
-    "faucet": "0.0.1",
-    "opn-cli": "^3.0.1",
-    "tape": "^4.0.0"
-  },
-  "scripts": {
-    "test": "make"
+    "eslint": "^2.13.1",
+    "eslint-config-airbnb-base": "^3.0.1",
+    "eslint-plugin-import": "^1.9.2",
+    "nyc": "^6.6.1",
+    "opn-cli": "^3.0.1"
   }
 }
